pwm.h and fan.h refactor
